Move forms from views file
