+++ reports/views.py
@@ -2,10 +2,9 @@
 from datetime import date
 
 # Django
-from django.views.generic import View
+from django.views import View
 from django.shortcuts import render
 from django.http import HttpResponse
-from django import forms
 from django.contrib.auth.decorators import login_required, user_passes_test
 from django.utils.decorators import method_decorator
 
@@ -13,32 +12,7 @@ from django.utils.decorators import method_decorator
 from reports.report_builder import XLSXReportBuilder
 
 from reports.export_builder import XLSXDataExportBuilder
-from reports.report_details import get_countries, get_regions
-
-
-class GlobalForm(forms.Form):
-    pass
-
-class CountryForm(forms.Form):
-    # Only show countries for which data has been submitted
-    COUNTRIES = get_countries()
-
-    country = forms.ChoiceField(
-        label='Country',
-        choices=COUNTRIES)
-
-    def filter_countries(self):
-        if self.cleaned_data['country'] == 'ALL':
-            return get_countries()
-        else:
-            return [(code, country) for code, country in get_countries() if code == self.cleaned_data['country']]
-
-class RegionForm(forms.Form):
-    REGIONS = get_regions()
-
-    region = forms.ChoiceField(
-        label='Region',
-        choices=REGIONS)
+from reports.forms import GlobalForm, CountryForm, RegionForm
 
 
 class ReportView(View):
@@ -62,10 +36,10 @@ class ReportView(View):
     def post(self, request, *args, **kwargs):
         if 'country_form' in request.POST:
             form = CountryForm(request.POST)
-            choice = [country for code, country in form.COUNTRIES if code == request.POST['country']][0]
+            choice = [country for code, country in form.get_form_countries() if code == request.POST['country']][0]
         elif 'region_form' in request.POST:
             form = RegionForm(request.POST)
-            choice = [region for id, region in form.REGIONS if id == int(request.POST['region'])][0]
+            choice = [region for id, region in form.get_form_regions() if id == int(request.POST['region'])][0]
         else:
             form = GlobalForm(request.POST)
             choice = 'Global'
